Update src/net/aisd/martin/frc2013/Controller.java
I added the rest of the buttons on an xbox controller. Since I didn't really know what the raw button number was, I just put the word number in place of it.
Also, some controllers register some form of sensitivity on the triggers (not pushed, half down, fully held down), this could be useful for the game.